ethicaldrm-main/ethicaldrm-main/ramijraj-develop-ethical-drm-kit-for-creators/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
imagehash>=4.3.1
flask>=2.3.0
requests>=2.31.0
telethon>=1.28.0
psutil>=5.9.0
cryptography>=41.0.0
pymongo>=4.5.0
pytest>=7.4.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-magic>=0.4.27
tqdm>=4.65.0
aiohttp>=3.8.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
google-generativeai
waitress
cryptography==45.0.7
distro==1.9.0
dnspython==2.8.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.1.2
frozenlist==1.7.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.2
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.6.4
numpy==2.2.6
openai==1.107.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.0
pyparsing==3.2.3
pytest==8.4.2
python-dotenv==1.1.1
python-magic==0.4.27
PyWavelets==1.9.0
requests==2.32.5
rsa==4.9.1
schedule==1.2.2
scipy==1.16.1
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8
Telethon==1.41.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.20.1